--- src/FluentValidation.Mvc5/FluentValidation.Mvc5.csproj
@@ -1,59 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net45</TargetFramework>
-    <AssemblyName>FluentValidation.Mvc</AssemblyName>
-    <RootNamespace>FluentValidation.Mvc</RootNamespace>
-    <PackageId>FluentValidation.Mvc5</PackageId>
-    <Company />
-    <Description>ASP.NET MVC 5 integration for FluentValidation</Description>
-    <PackageReleaseNotes>https://github.com/JeremySkinner/FluentValidation/blob/master/Changelog.txt</PackageReleaseNotes>
-    <!-- Signing is only performed as part of the release build -->
-    <SignAssembly Condition="'$(AssemblyOriginatorKeyFile)'!=''">True</SignAssembly>
-    <DelaySign>False</DelaySign>
-    <DocumentationFile Condition="'$(Configuration)'=='Release'">bin\$(Configuration)\$(TargetFramework)\FluentValidation.Mvc.xml</DocumentationFile>
-	  <PackageReleaseNotes>
-Changes in 8.6.0:
-* AddToModelState no longer includes a dot if PropertyName is null
-
-Changes in 8.4.0:
-* Added simplified version of several messages across all translations to aid with ASP.NET Clientside validation.
-
-Changes in 8.1.3:
-* Resolved issue where clientside validation was incorrectly generating metadata for rules with shared conditions.
-
-Changes in 8.1:
-* Added the Skip property to the CustomizeValidatorAttribute to bring parity with the ASP.NET Core integration.
-
-Changes in 7.5:
-* Better handling of clientside messages when using custom overriden language resources
-* Added flag in the ValidationContext's RootDataContext when a validator is invoked by the automatic MVC/WebApi integration
-
-	Changes in 7.4:
-	* Added SetRulesetForClientsideMessages method as an alternative to using the RulesetForClientSideMessagesAttribute
-	* Updated the MVC5 AddToModelState extension not to overwrite existing attempted values
-
-	Full release notes can be found at https://github.com/JeremySkinner/FluentValidation/blob/master/Changelog.txt
-		</PackageReleaseNotes>
-      <DebugType>embedded</DebugType>
-	</PropertyGroup>
-
-  <ItemGroup>
-    <Compile Include="..\CommonAssemblyInfo.cs" Link="CommonAssemblyInfo.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNet.Mvc" Version="5.2.0" />
-  </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\FluentValidation\FluentValidation.csproj" />
-      <ProjectReference Include="..\FluentValidation.ValidatorAttribute\FluentValidation.ValidatorAttribute.csproj" />
-    </ItemGroup>
-
-  <ItemGroup>
-    <Reference Include="System.ComponentModel.DataAnnotations" />
-    <Reference Include="System.Web" />
-  </ItemGroup>
-
-  <Import Project="../netfx.props" />
-</Project>
