+++ src/FluentValidation.Mvc5/README.txt
@@ -1,0 +1,1 @@
+ASP.NET MVC 5 integration has been moved to https://github.com/JeremySkinner/FluentValidation-LegacyWeb
