--- src/FluentValidation.ValidatorAttribute/FluentValidation.ValidatorAttribute.csproj
@@ -1,17 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworks>netstandard1.1;netstandard2.0;net45</TargetFrameworks>
-    <Description>A validation library for .NET that uses a fluent interface to construct strongly-typed validation rules.</Description>
-    <!-- Signing is only performed as part of the release build -->
-    <SignAssembly Condition="'$(AssemblyOriginatorKeyFile)'!=''">True</SignAssembly>
-    <DelaySign>False</DelaySign>
-    <DebugType>embedded</DebugType>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="..\CommonAssemblyInfo.cs" Link="CommonAssemblyInfo.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\FluentValidation\FluentValidation.csproj" />
-  </ItemGroup>
-  <Import Project="../netfx.props" />
-</Project>
