--- src/FluentValidation.WebApi/FluentValidation.WebApi.csproj
@@ -1,41 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net45</TargetFramework>
-    <PackageId>FluentValidation.WebApi</PackageId>
-    <Description>ASP.NET WebApi integration for FluentValidation</Description>
-    <!-- Signing is only performed as part of the release build -->
-    <SignAssembly Condition="'$(AssemblyOriginatorKeyFile)'!=''">True</SignAssembly>
-    <DelaySign>False</DelaySign>
-    <DebugType>embedded</DebugType>
-    <DocumentationFile Condition="'$(Configuration)'=='Release'">bin\$(Configuration)\$(TargetFramework)\FluentValidation.WebApi.xml</DocumentationFile>
-    <PackageReleaseNotes>
-Changes in 8.6.0:
-* AddToModelState no longer includes a dot if PropertyName is null
-
-Changes in 8.2.3:
-* Fix issue in WebApi integration where DataAnnotations validators could generate duplicate messages when in the same project as FluentValidation.
-
-Changes in 8.1.0:
-* Added Skip property on CustomizeValidatorAttribute for consistency with ASP.NET Core.
-
-Changes in 7.5.1:
-* Added support for CustomizeValidatorAttribute for consistency with MVC5 and AspNetCore.
-
-Changes in 7.4:
-* Support for WebApi 5.2.4 and improved integration with WebApi's validation pipeline.
-
-Full release notes can be found at https://github.com/JeremySkinner/FluentValidation/blob/master/Changelog.txt
-</PackageReleaseNotes>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="..\CommonAssemblyInfo.cs" Link="CommonAssemblyInfo.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNet.WebApi.Core" Version="5.2.4" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\FluentValidation\FluentValidation.csproj" />
-    <ProjectReference Include="..\FluentValidation.ValidatorAttribute\FluentValidation.ValidatorAttribute.csproj" />
-  </ItemGroup>
-  <Import Project="../netfx.props" />
-</Project>
