+++ src/FluentValidation.WebApi/README.txt
@@ -1,0 +1,1 @@
+WebApi 2 integration has been moved to https://github.com/JeremySkinner/FluentValidation-LegacyWeb
